requirements.txt
dj-database-url==0.5.0
Django==2.2.2
flake8==3.7.8
gunicorn==19.9.0
ipdb==0.12
psycopg2-binary==2.8.3
whitenoise==4.1.2
requests==2.22.0
django-import-export==1.2.0